pyproject.toml
[project]
name = "cineinfo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.13",
    "google-genai>=1.21.1",
    "pydantic-settings>=2.10.0",
    "uvicorn>=0.34.3",
]
